+++ CHANGES.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - iOS: Fix for #216: "when 'connecting' there is no 'disconnect' button"
+- macOS: Support for "on demand" preference added #220
 - macOS: Fix UI not updating correctly after removing provider #213
 - macOS: Don't tell users about them pressing cancel, they know already
 - macOS: Added popup for uninstall instructions older app #212